Revert "Refactored "routes" folder"
